+++ Assets/Scripts/Game/Score.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-// Scoring system: the longer Player survive the more score gained
+// COntroller for Scoring system: the longer Player survive the more score gained
+// - Also Control High Score Setting and Saving
 public class Score: MonoBehaviour {
     public float timeStep = 1f;
     public Text scoreText;
@@ -11,7 +12,27 @@ public class Score: MonoBehaviour {
     Player player;
     float currentStep = 0f;
 
+    public int CurrentScore {
+        get {
+            return score;
+        }
+    }
+
+    // Using PlayerPrefs for Storing and Saving Highscore on Game Over
+    public bool HighScoreCheck() {
+        int highScore = PlayerPrefs.GetInt("High Score", 0);
+        if (score > highScore) {
+            PlayerPrefs.SetInt("High Score", score);
+            Debug.Log("New High Score!");
+            return true;
+        } else {
+            Debug.Log("No High Score");
+            return false;
+        }
+    }
+
     void Start() {
+        score = 0;
         player = GameObject.FindWithTag("Player").GetComponent<Player>();
         scoreText.text = score.ToString();
     }